+++ src/app/(tools)/gear-solver/GearSolverWrapper.tsx
@@ -1,0 +1,8 @@
+'use client';
+import dynamic from 'next/dynamic';
+
+const GearSolverClient = dynamic(() => import('./gearSolverClient'), { ssr: false });
+
+export default function GearSolverWrapper() {
+  return <GearSolverClient />;
+}